Enable closures with model arguments.
An Exception is thrown if you use the Model typehint / type as an argument(s) in your repository methods, a pretty common practice. `Serialization of 'Closure' is not allowed` There may be a more clever way to support more types of objects, but this was fast, easy and worked for my purposes.